+++ Demo2/MainWindow.xaml.cs
@@ -1,0 +1,15 @@
+﻿// -----------------------------------------------------------------------------------------------------------------
+// <copyright file="MainWindow.xaml.cs" company="my-libraries">
+//     Copyright (c) David Wendland. All rights reserved.
+// </copyright>
+// -----------------------------------------------------------------------------------------------------------------
+
+namespace Demo2;
+
+public partial class MainWindow
+{
+    public MainWindow()
+    {
+        InitializeComponent();
+    }
+}