+++ Demo2/Properties/AssemblyInfo.cs
@@ -1,0 +1,9 @@
+// -----------------------------------------------------------------------------------------------------------------
+// <copyright file="AssemblyInfo.cs" company="my-libraries">
+//     Copyright (c) David Wendland. All rights reserved.
+// </copyright>
+// -----------------------------------------------------------------------------------------------------------------
+
+using System.Windows;
+
+[assembly: ThemeInfo(ResourceDictionaryLocation.None, ResourceDictionaryLocation.SourceAssembly)]